#27: end2end galerkin rom support the dryrun option

+++ end-to-end-roms/wf_galerkin.py
@@ -22,50 +22,53 @@ def _validate_wf_galerkin_section(wfDic, customModule):
 
 def compute_hypreducer_matrix_operator(outDir, offlineRomDir, \
                                        numStateModes, sampleMeshDir, \
-                                       numDofsPerCell):
-  # load my full phi
+                                       numDofsPerCell, dryRun):
+
   myFullPhiFile = offlineRomDir + "/state_left_singular_vectors.bin"
-  myFullPhi     = load_pod_basis(myFullPhiFile)[:,0:numStateModes]
-
-  # indexing info
-  myFile2      = sampleMeshDir + "/sample_mesh_gids.txt"
-  mySmMeshGids = np.loadtxt(myFile2, dtype=int)
-  mySmCount    = len(mySmMeshGids)
-
-  K = numStateModes*3 + 1
-  if mySmCount*numDofsPerCell < K:
-    print("Cannot have K > mySmCount*numDofsPerCell, using K")
-    K = mySmCount*numDofsPerCell - 1
-
-  # K should be larger than numStateModes
-  if K < numStateModes:
-    print("Cannot have K < myNumStatePodModes, using K")
-    K = numStateModes + 1
-
+  myFile2 = sampleMeshDir + "/sample_mesh_gids.txt"
   fullRhsPodFile = offlineRomDir + "/rhs_left_singular_vectors.bin"
-  theta = load_pod_basis(fullRhsPodFile)[:,0:K]
-  slicedTheta = np.zeros((mySmCount*numDofsPerCell, theta.shape[1]), order='F')
-  for j in range(numDofsPerCell):
-    slicedTheta[j::numDofsPerCell, :] = theta[numDofsPerCell*mySmMeshGids + j, :]
+  K = numStateModes*3 + 1
 
-  A = myFullPhi.transpose() @ theta
-  hypRedOp = A @ scipyla.pinv(slicedTheta)
-  print(" hyperReductionOperatorShape = {}".format(hypRedOp.shape))
+  if not dryRun:
+    # load my full phi
+    myFullPhi     = load_pod_basis(myFullPhiFile)[:,0:numStateModes]
 
-  numRows = np.int64(hypRedOp.shape[1])
-  numCols = np.int64(hypRedOp.shape[0])
-  print("numRows = ", numRows)
-  print("numCols = ", numCols)
-  fileo = open(outDir+'/hyperReductionOperator.bin', "wb")
-  np.array([numRows]).tofile(fileo)
-  np.array([numCols]).tofile(fileo)
-  hypRedOp.tofile(fileo)
-  fileo.close()
-  #np.savetxt(outDir+'/hyperReductionOperator.txt', hypRedOp.T)
+    # indexing info
+    mySmMeshGids = np.loadtxt(myFile2, dtype=int)
+    mySmCount    = len(mySmMeshGids)
+
+    if mySmCount*numDofsPerCell < K:
+      print("Cannot have K > mySmCount*numDofsPerCell, using K")
+      K = mySmCount*numDofsPerCell - 1
+
+    # K should be larger than numStateModes
+    if K < numStateModes:
+      print("Cannot have K < myNumStatePodModes, using K")
+      K = numStateModes + 1
+
+    theta = load_pod_basis(fullRhsPodFile)[:,0:K]
+    slicedTheta = np.zeros((mySmCount*numDofsPerCell, theta.shape[1]), order='F')
+    for j in range(numDofsPerCell):
+      slicedTheta[j::numDofsPerCell, :] = theta[numDofsPerCell*mySmMeshGids + j, :]
+
+    A = myFullPhi.transpose() @ theta
+    hypRedOp = A @ scipyla.pinv(slicedTheta)
+    print(" hyperReductionOperatorShape = {}".format(hypRedOp.shape))
+
+    numRows = np.int64(hypRedOp.shape[1])
+    numCols = np.int64(hypRedOp.shape[0])
+    print("numRows = ", numRows)
+    print("numCols = ", numCols)
+    fileo = open(outDir+'/hyperReductionOperator.bin', "wb")
+    np.array([numRows]).tofile(fileo)
+    np.array([numCols]).tofile(fileo)
+    hypRedOp.tofile(fileo)
+    fileo.close()
+    #np.savetxt(outDir+'/hyperReductionOperator.txt', hypRedOp.T)
 
 
 def _run_single_rom(romDir, offlineRomDir, numModes, inputsDic, \
-                    fomRunDir, algoString, \
+                    fomRunDir, algoString, dryRun,
                     sampleMeshDir = None, hypRedOpDir = None):
 
   if os.path.exists(romDir):
@@ -85,15 +88,16 @@ def _run_single_rom(romDir, offlineRomDir, numModes, inputsDic, \
     fullMeshPodFile = offlineRomDir +"/state_left_singular_vectors.bin"
     romSpecifics['fullMeshPodFile'] = fullMeshPodFile
 
-    # compute rom initial state using the basis
-    myPhi = load_pod_basis(fullMeshPodFile)[:, 0:numModes]
-    fomIc = load_fom_initial_condition(fomRunDir+"/initial_state.bin")
-    romIc = np.dot(myPhi.transpose(), fomIc)
-    romInitStateFile = romDir+"/rom_initial_state.txt"
-    np.savetxt(romInitStateFile, romIc)
+    romInitStateFile = romDir + "/rom_initial_state.txt"
+    if not dryRun:
+      # compute rom initial state using the basis
+      myPhi = load_pod_basis(fullMeshPodFile)[:, 0:numModes]
+      fomIc = load_fom_initial_condition(fomRunDir+"/initial_state.bin")
+      romIc = np.dot(myPhi.transpose(), fomIc)
+      np.savetxt(romInitStateFile, romIc)
 
-    # deal with affine shift
-    np.savetxt(romDir+"/optional_affine_shift.txt", np.zeros(len(fomIc)))
+      # deal with affine shift
+      np.savetxt(romDir+"/optional_affine_shift.txt", np.zeros(len(fomIc)))
     romSpecifics['affineShiftFile'] = romDir +"/optional_affine_shift.txt"
     romSpecifics['isAffine'] = False
 
@@ -118,7 +122,7 @@ def _run_single_rom(romDir, offlineRomDir, numModes, inputsDic, \
       link_if_needed_and_run_exe(romDir,  exeDirFullPath, cppExecutableName)
 
 
-def _main_default_impl(workDir, romDic):
+def _main_default_impl(workDir, romDic, dryRun=False):
   # location of the offline stuff
   offlineRomDir = workDir + "/offline_rom"
 
@@ -131,9 +135,12 @@ def _main_default_impl(workDir, romDic):
 
     if (truncPolicy.lower() == "energybased"):
       for energy in truncValues:
-        singValues = np.loadtxt(offlineRomDir+'/state_singular_values.txt')
-        numModes = compute_cumulative_energy(singValues, energy)
-        print(numModes)
+        if dryRun:
+          numModes = 22
+        else:
+          singValues = np.loadtxt(offlineRomDir+'/state_singular_values.txt')
+          numModes = compute_cumulative_energy(singValues, energy)
+          print(numModes)
 
         #
         # for a given num of modes, find all FOM test runs
@@ -153,10 +160,17 @@ def _main_default_impl(workDir, romDic):
           romDir += "_"+str(energy)
           romDir += "_runid_"+str(runId)
           _run_single_rom(romDir, offlineRomDir, numModes, \
-                          fomInputs, fomTestDir, "defaultGalerkin")
+                          fomInputs, fomTestDir, "defaultGalerkin", dryRun)
 
 # -------------------------------------------------------------------
-def _main_hyperreduced_impl(workDir, romDic, numDofsPerCell):
+def string_identifier_from_sample_mesh_dir(sampleMeshDir):
+  if "sample_mesh_random" in sampleMeshDir:
+    return "sample_mesh_random_"+sampleMeshDir[-5:]
+  elif "sample_mesh_psampling" in sampleMeshDir:
+    return "sample_mesh_psampling_"+sampleMeshDir[-5:]
+
+# -------------------------------------------------------------------
+def _main_hyperreduced_impl(workDir, romDic, numDofsPerCell, dryRun=False):
   # location of the offline stuff
   offlineRomDir = workDir + "/offline_rom"
 
@@ -173,9 +187,12 @@ def _main_hyperreduced_impl(workDir, romDic, numDofsPerCell):
       print (truncPolicy, truncValues)
       if (truncPolicy.lower() == "energybased"):
         for energy in truncValues:
-          singValues = np.loadtxt(offlineRomDir+'/state_singular_values.txt')
-          numModes = compute_cumulative_energy(singValues, energy)
-          print(numModes)
+          if dryRun:
+            numModes = 22
+          else:
+            singValues = np.loadtxt(offlineRomDir+'/state_singular_values.txt')
+            numModes = compute_cumulative_energy(singValues, energy)
+            print(numModes)
 
           s1 = workDir + "/hyperreducer"
           if energy != None:
@@ -192,7 +209,7 @@ def _main_hyperreduced_impl(workDir, romDic, numDofsPerCell):
 
           compute_hypreducer_matrix_operator(hypRedPath, offlineRomDir, \
                                              numModes, sampleMeshDir, \
-                                             numDofsPerCell)
+                                             numDofsPerCell, dryRun)
 
           for fomTestDir in find_all_fom_test_dirs(workDir):
             # read the yaml file used for that FOM run
@@ -209,7 +226,7 @@ def _main_hyperreduced_impl(workDir, romDic, numDofsPerCell):
             romDir += "_runid_"+str(runId)
             print(romDir)
             _run_single_rom(romDir, offlineRomDir, numModes, fomInputs,\
-                            fomTestDir, "hyperreducedGalerkin",\
+                            fomTestDir, "hyperreducedGalerkin", dryRun,
                             sampleMeshDir = sampleMeshDir,\
                             hypRedOpDir = hypRedPath)
 
@@ -268,7 +285,6 @@ if __name__== "__main__":
 
   # figure out which galerkin we want and run
   if romDic['algorithm'].lower() == "defaultgalerkin":
-    _main_default_impl(workDirFullPath, romDic)
-
+    _main_default_impl(workDirFullPath, romDic, dryRun)
   elif romDic['algorithm'].lower() == "hyperreducedgalerkin":
-    _main_hyperreduced_impl(workDirFullPath, romDic, customModule.numDofsPerCell)
+    _main_hyperreduced_impl(workDirFullPath, romDic, customModule.numDofsPerCell, dryRun)
